+++ src/core/ModuleFactory.cpp
@@ -13,6 +13,7 @@
  */
 
 #include "ModuleFactory.h"
+#include "Logger.h"
 
 // Implementation of the getInstance() method
 ModuleFactory& ModuleFactory::getInstance() {
@@ -22,9 +23,7 @@ ModuleFactory& ModuleFactory::getInstance() {
 
 // Registers a new module type and its creation function
 void ModuleFactory::registerModule(const std::string& type, ModuleCreationFunction func) {
-    Serial.print("ModuleFactory: Registering type '");
-    Serial.print(type.c_str());
-    Serial.println("'");
+    NEXTINO_CORE_LOG(LogLevel::Debug, "ModFactory", "Registering type '%s'", type.c_str());
     registry[type] = func;
 }
 
@@ -32,15 +31,11 @@ void ModuleFactory::registerModule(const std::string& type, ModuleCreationFuncti
 BaseModule* ModuleFactory::createModule(const std::string& type, const JsonObject& config) {
     // Check if the type is registered
     if (registry.find(type) != registry.end()) {
-        Serial.print("ModuleFactory: Creating module of type '");
-        Serial.print(type.c_str());
-        Serial.println("'...");
+        NEXTINO_CORE_LOG(LogLevel::Debug, "ModFactory", "Creating module of type '%s'...", type.c_str());
         // Call the stored function and pass the configuration
         return registry[type](config);
     }
 
-    Serial.print("ModuleFactory: ERROR - Unknown module type '");
-    Serial.print(type.c_str());
-    Serial.println("'");
+    NEXTINO_CORE_LOG(LogLevel::Error, "ModFactory", "Unknown module type '%s'", type.c_str());
     return nullptr; // If the type is not found, return nullptr
 }